Massive refactor
-Every resource is now allocated once with the max
-Vulkan and Dx12 allocated the buffers and copy the data later
-Dx12 works, Vulkan TBD
-Moved multiple header files from .h files to .cpp

+++ src/Core/DbLog/blitLogger.h
@@ -69,7 +69,7 @@ namespace BlitzenCore
 
 #else
         
-        if (BlitzenCore::BLIT_CHECK_FAIL(level))
+        if (BlitzenCore::BLIT_CHECK_FAIL((int64_t)level))
         {
             BlitzenPlatform::PlatformConsoleError(outMessage2, level);
         }
